refactor(agentos): rename — not a Service, just a function in scheduling/ (#11858)
Operator: "this is not a service. it is a function." Service/Coordinator
suffixes imply state + methods + lifecycle; this file holds one pure
trigger projection. Files describe what they are.

Renames:
- ai/daemons/orchestrator/services/DreamCoordinatorService.mjs
  -> ai/daemons/orchestrator/scheduling/dream.mjs
- test/.../services/DreamCoordinatorService.spec.mjs
  -> test/.../scheduling/dream.spec.mjs

New folder scheduling/ (sibling to services/) holds per-task due-trigger
projections. services/ keeps things that are genuine services. Example:
SummarizationCoordinatorService calls getUnreadSunsetHandoversFn(db) and
attaches onSuccess hooks; that one EARNED the Service suffix. DreamService
/ SwarmHeartbeatService / GoldenPathSynthesizer stay in services/ as
execution-side responsibilities.

JSDoc + describe title updated. No more Coordinator/Service in this file
or its spec. 4/4 spec tests pass at new location.

Sub 18's registry will consume via:
  import {getDueTask as dreamGetDueTask} from './scheduling/dream.mjs'

Downstream churn flagged: Discussion #11857 body + tickets #11859/11860
still use ...CoordinatorService naming. Will update when filing Sub 16 /
Sub 17 work (same shape: scheduling/<task>.mjs). Sub 19 #11861 stays
Service-named since MaintenanceBackpressureService IS a genuine service
(state + multiple visibly-separate surfaces + lease delegation).
